pyproject.toml
[project]
name = "storesim"
version = "0.1.0"
description = "Store product similarity search using CLIP ResNet50"
authors = [{name = "StoreSim"}]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ftfy>=6.0.0",
    "regex>=2023.0.0",
    "tqdm>=4.65.0",
    "Pillow>=9.0.0",
    "numpy>=1.24.0",
    "open-clip-torch>=2.24.0",
    "datasets>=2.14.0",
    "huggingface-hub>=0.20.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pymilvus>=2.4.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "dvc>=3.30.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
storesim-ui = "app.main:_run"

[tool.hatch.build.targets.wheel]
packages = ["storesim", "app"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=storesim --cov-report=term-missing"

[tool.coverage.run]
source = ["storesim"]
omit = ["tests/*"]

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "ANN", # flake8-annotations (type hints)
    "RUF", # ruff-specific rules
]
ignore = [
    "ANN401",  # allow Any in return types
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]  # don't require type annotations in tests

[tool.ruff.lint.isort]
known-first-party = ["storesim"]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests/"]
